+++ compiler-core/src/format.rs
@@ -255,10 +255,11 @@ impl<'comments> Formatter<'comments> {
                     ".{".to_doc().append(unqualified).append("}")
                 };
                 let doc = docvec!["import ", module.as_str(), second];
-                if let Some(as_name) = as_name {
-                    doc.append(" as ").append(&as_name.name)
-                } else {
-                    doc
+                match as_name {
+                    ImportName::Alias(_, name) | ImportName::Discarded(_, name) => {
+                        doc.append(" as ").append(name)
+                    }
+                    ImportName::Original(_, _) => doc,
                 }
             }